Termino das telas de perfilVaga e ARRUMAR O tecnologias com o CCheckComboBox que nao pegou na minha maquina! Gabi arruma ai que tu consegue importar esse treco da tela Criar Vaga

+++ FoundIt/src/main/java/gui/TelaEditarVagaController.java
@@ -119,7 +119,7 @@ public class TelaEditarVagaController implements Initializable {
             String descricaoVaga = editarDescricao.getText();
             String nivelVaga = editarNivel.getText();
             String localVaga = editarLocal.getText();
-            int salarioVaga = (int) Integer.parseInt(editarSalario.getText());
+            double salarioVaga = (double) Double.parseDouble(editarSalario.getText());
             TipoContrato contratoVaga = editarContrato.getSelectionModel().getSelectedItem();
 
             ObservableList<Tecnologias> tecnologiaVaga = editarTecnologias.getCheckModel().getCheckedItems();
